UI changes to Command UI part 2
