--- kfbot-refactor-phase2/src/main/java/com/twitchbotx/gui/CommandEditor.fxml
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.control.MenuButton?>
-<?import javafx.scene.control.MenuItem?>
-<?import javafx.scene.control.RadioButton?>
-<?import javafx.scene.control.TextArea?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.control.ToggleGroup?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.twitchbotx.gui.controllers.CommandEditorController">
-   <children>
-      <Button layoutX="10.0" layoutY="361.0" mnemonicParsing="false" onAction="#dash" text="Dashboard" />
-      <TextField fx:id="commandToEdit" layoutX="302.0" layoutY="44.0" />
-      <Label layoutX="236.0" layoutY="7.0" text="Edit a command">
-         <font>
-            <Font size="18.0" />
-         </font>
-      </Label>
-      <Button fx:id="submitChanges" layoutX="465.0" layoutY="123.0" mnemonicParsing="false" onAction="#submitChanges" prefHeight="25.0" prefWidth="124.0" text="Submit Changes" />
-      <Button fx:id="deleteCommand" layoutX="460.0" layoutY="44.0" mnemonicParsing="false" onAction="#deleteCommand" prefHeight="25.0" prefWidth="129.0" text="Delete Command" />
-      <Label layoutX="214.0" layoutY="200.0" text="Create new command">
-         <font>
-            <Font size="18.0" />
-         </font>
-      </Label>
-      <TextField fx:id="newCommandName" layoutX="73.0" layoutY="241.0" prefHeight="25.0" prefWidth="107.0" promptText="!command" />
-      <TextField fx:id="newCommandAuth" layoutX="73.0" layoutY="277.0" prefHeight="25.0" prefWidth="133.0" promptText="(-s +m -a +username)" />
-      <Label layoutX="29.0" layoutY="245.0" text="Name:" />
-      <Label layoutX="33.0" layoutY="281.0" text="Auth:" />
-      <Label layoutX="4.0" layoutY="321.0" text="Cooldown:" />
-      <TextField fx:id="newCommandCooldown" alignment="CENTER" layoutX="73.0" layoutY="317.0" prefHeight="25.0" prefWidth="51.0" text="30" />
-      <Label layoutX="135.0" layoutY="321.0" text="seconds" />
-      <Label layoutX="197.0" layoutY="93.0" text="What to edit">
-         <font>
-            <Font size="14.0" />
-         </font>
-      </Label>
-      <MenuButton fx:id="attributesMenu" layoutX="194.0" layoutY="123.0" mnemonicParsing="false" text="Attributes">
-        <items>
-          <MenuItem fx:id="attributeAuth" mnemonicParsing="false" onAction="#setAttribute" text="Auth" />
-          <MenuItem fx:id="attributeCooldown" mnemonicParsing="false" onAction="#setAttribute" text="Cooldown" />
-            <MenuItem fx:id="attributeEnabled" mnemonicParsing="false" onAction="#setAttribute" text="Disabled" />
-            <MenuItem fx:id="attributeInitialDelay" mnemonicParsing="false" onAction="#setAttribute" text="Initial Delay" />
-            <MenuItem fx:id="attributeInterval" mnemonicParsing="false" onAction="#setAttribute" text="Interval" />
-            <MenuItem mnemonicParsing="false" onAction="#setAttribute" text="Repeating" />
-            <MenuItem fx:id="attributeSound" mnemonicParsing="false" onAction="#setAttribute" text="Sound" />
-            <MenuItem fx:id="attributeMessage" mnemonicParsing="false" onAction="#setAttribute" text="Message" />
-        </items>
-      </MenuButton>
-      <ListView fx:id="commandList" layoutX="32.0" layoutY="34.0" prefHeight="158.0" prefWidth="149.0" />
-      <Label layoutX="58.0" layoutY="12.0" text="Select a command" />
-      <Button layoutX="197.0" layoutY="44.0" mnemonicParsing="false" onAction="#chooseCommand" text="Select" />
-      <Label layoutX="253.0" layoutY="48.0" text="-&gt;" />
-      <Label fx:id="editValueLabel" layoutX="333.0" layoutY="93.0" text="New value">
-         <font>
-            <Font size="14.0" />
-         </font>
-      </Label>
-      <Label layoutX="217.0" layoutY="245.0" text="Repeating:" />
-      <Label layoutX="213.0" layoutY="289.0" text="Initial Delay:" />
-      <Label layoutX="216.0" layoutY="321.0" text="Interval:" />
-      <Button layoutX="455.0" layoutY="361.0" mnemonicParsing="false" onAction="#newCommand" text="Add New Command" />
-      <Label layoutX="407.0" layoutY="245.0" text="Message:" />
-      <TextField fx:id="newCommandInitDelay" alignment="CENTER" layoutX="292.0" layoutY="285.0" prefHeight="25.0" prefWidth="35.0" text="0" />
-      <Label layoutX="344.0" layoutY="289.0" text="seconds" />
-      <RadioButton fx:id="repeatYes" layoutX="290.0" layoutY="236.0" mnemonicParsing="false" text="True">
-         <toggleGroup>
-            <ToggleGroup fx:id="repeating" />
-         </toggleGroup>
-      </RadioButton>
-      <RadioButton fx:id="repeatNo" layoutX="291.0" layoutY="256.0" mnemonicParsing="false" selected="true" text="False" toggleGroup="$repeating" />
-      <Label layoutX="344.0" layoutY="321.0" text="seconds" />
-      <TextField fx:id="newCommandInterval" alignment="CENTER" layoutX="272.0" layoutY="317.0" prefHeight="25.0" prefWidth="57.0" text="1800" />
-      <Label fx:id="submitStatus" layoutX="465.0" layoutY="164.0" prefHeight="56.0" prefWidth="103.0" wrapText="true" />
-      <Label fx:id="cmdStatus" alignment="CENTER" contentDisplay="CENTER" layoutX="421.0" layoutY="182.0" prefHeight="64.0" prefWidth="149.0" textAlignment="CENTER" textFill="#f80707" wrapText="true" />
-      <Label layoutX="114.0" layoutY="365.0" text="Sound file (optional):" />
-      <TextField fx:id="newCommandSound" layoutX="253.0" layoutY="361.0" prefHeight="25.0" prefWidth="149.0" promptText="filename.mp3" />
-      <TextArea fx:id="newValueText" layoutX="302.0" layoutY="123.0" maxHeight="75.0" maxWidth="149.0" minHeight="25.0" minWidth="149.0" prefHeight="25.0" prefWidth="149.0" wrapText="true" />
-      <TextArea fx:id="newCommandMessage" layoutX="409.0" layoutY="269.0" prefHeight="82.0" prefWidth="174.0" wrapText="true" />
-      <Button layoutX="20.0" layoutY="200.0" mnemonicParsing="false" onAction="#repeatCommandManager" prefHeight="27.0" prefWidth="174.0" text="Repeating Commands" textAlignment="CENTER" wrapText="true" />
-      <Label fx:id="deleteConfirmText" layoutX="442.0" layoutY="71.0" prefHeight="50.0" prefWidth="149.0" textAlignment="CENTER" textFill="RED" wrapText="true" />
-      <Button layoutX="517.0" layoutY="241.0" mnemonicParsing="false" onAction="#clearNewCommand" text="Clear New" />
-   </children>
-</AnchorPane>
