--- kfbot-refactor-phase2/src/main/java/com/twitchbotx/gui/TimedCommands.fxml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.twitchbotx.gui.controllers.TimedCommandsController">
-   <children>
-      <Button layoutX="22.0" layoutY="361.0" mnemonicParsing="false" onAction="#dash" text="Dashboard" />
-      <Button layoutX="158.0" layoutY="361.0" mnemonicParsing="false" onAction="#editCommands" text="Command Editor" />
-      <ScrollPane fx:id="commandList" layoutX="30.0" layoutY="48.0" prefHeight="276.0" prefWidth="171.0" />
-      <Label layoutX="46.0" layoutY="14.0" text="Available Commands">
-         <font>
-            <Font name="System Bold" size="14.0" />
-         </font>
-      </Label>
-      <Label layoutX="434.0" layoutY="14.0" text="Options">
-         <font>
-            <Font name="System Bold" size="14.0" />
-         </font>
-      </Label>
-      <Label layoutX="271.0" layoutY="59.0" text="Name:" />
-      <Label layoutX="263.0" layoutY="127.0" text="Interval:" />
-      <Label layoutX="241.0" layoutY="158.0" text="Initial Delay:" />
-      <Label layoutX="257.0" layoutY="192.0" text="Message:" />
-      <Button layoutX="540.0" layoutY="359.0" mnemonicParsing="false" onAction="#startCommand" text="Start" />
-      <Button layoutX="341.0" layoutY="359.0" mnemonicParsing="false" onAction="#stopCommand" text="Stop" />
-      <Label fx:id="commandNameText" alignment="CENTER" contentDisplay="CENTER" layoutX="380.0" layoutY="58.0" prefHeight="19.0" prefWidth="160.0" style="-fx-border-color: grey;" textAlignment="CENTER" />
-      <Label fx:id="commandInterval" alignment="CENTER" contentDisplay="CENTER" layoutX="380.0" layoutY="126.0" prefHeight="19.0" prefWidth="160.0" style="-fx-border-color: grey;" textAlignment="CENTER" />
-      <Label fx:id="commandInitDelay" alignment="CENTER" contentDisplay="CENTER" layoutX="380.0" layoutY="157.0" prefHeight="19.0" prefWidth="160.0" style="-fx-border-color: grey;" textAlignment="CENTER" />
-      <Label fx:id="commandMessage" alignment="CENTER" contentDisplay="TOP" layoutX="342.0" layoutY="193.0" prefHeight="151.0" prefWidth="242.0" style="-fx-border-color: grey; -fx-label-padding: 2;" textAlignment="CENTER" wrapText="true" />
-      <Label layoutX="249.0" layoutY="93.0" text="Repeating:" />
-      <Label fx:id="commandRepeating" alignment="CENTER" contentDisplay="CENTER" layoutX="380.0" layoutY="92.0" prefHeight="19.0" prefWidth="160.0" style="-fx-border-color: grey;" textAlignment="CENTER" />
-      <Label fx:id="confirmationLabel" alignment="CENTER" contentDisplay="CENTER" layoutX="386.0" layoutY="353.0" prefHeight="38.0" prefWidth="152.0" textAlignment="CENTER" wrapText="true">
-         <font>
-            <Font size="11.0" />
-         </font></Label>
-   </children>
-</AnchorPane>
